# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Secure marketplace for trading and auctioning items in Aethoria.\nCommon and Rare items trade via fixed-price listings. Legendary items trade via auctions.",
        "title": "Market Dragon API",
        "contact": {
            "name": "Market Dragon Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auctions/{id}": {
            "get": {
                "description": "Get one auction and its current highest bid (if any).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auctions"
                ],
                "summary": "Get an auction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Auction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuctionDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuctionDetailResponse": {
            "type": "object",
            "properties": {
                "auction": {
                    "$ref": "#/definitions/dto.AuctionResponse"
                },
                "highest_bid": {
                    "$ref": "#/definitions/dto.BidResponse"
                }
            }
        },
...
        "dto.AuctionResponse": {
            "type": "object",
            "properties": {
                "ends_at": {
                    "type": "string",
                    "example": "2026-01-02T00:00:00Z"
                },
                "highest_bid_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "item_id": {
                    "type": "integer",
                    "example": 3
                },
                "seller_guild_id": {
                    "type": "integer",
                    "example": 3
                },
                "starts_at": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "winner_guild_id": {
                    "type": "integer"
                }
            }
        },
        "dto.BidResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1200
                },
                "auction_id": {
                    "type": "integer",
                    "example": 1
                },
                "bidder_guild_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "item not found"
                }
            }
        },
...
    },
...
}